fix: fix codegen command
