+++ apps/journeys-admin/pages/templates/index.tsx
@@ -17,7 +17,7 @@ import {
 import { GetMe } from '../../__generated__/GetMe'
 import { GetTags } from '../../__generated__/GetTags'
 import { PageWrapper } from '../../src/components/PageWrapper'
-import { GET_ME } from '@core/admin/ui/NavigationDrawer/UserNavigation'
+import { GET_ME } from 'apps/journeys-admin/src/components/PageWrapper/UserNavigation'
 import { TemplateGallery } from '../../src/components/TemplateGallery'
 import { initAndAuthApp } from '../../src/libs/initAndAuthApp'
 import { GET_JOURNEYS } from '../../src/libs/useJourneysQuery/useJourneysQuery'
