+++ apps/journeys-admin/src/components/PageWrapper/UserNavigation/ImpersonateDialog/ImpersonateDialog.tsx
@@ -14,7 +14,7 @@ import { Dialog } from '@core/shared/ui/Dialog'
 import {
   UserImpersonate,
   UserImpersonateVariables
-} from '../../../../../../../../apps/journeys-admin/__generated__/UserImpersonate'
+} from '../../../../../__generated__/UserImpersonate'
 
 export const USER_IMPERSONATE = gql`
   mutation UserImpersonate($email: String!) {
