+++ apps/journeys-admin/src/components/PageWrapper/UserNavigation/UserNavigation.tsx
@@ -15,15 +15,12 @@ import { MouseEvent, ReactElement, useEffect, useState } from 'react'
 import BoxIcon from '@core/shared/ui/icons/Box'
 import UserProfile3Icon from '@core/shared/ui/icons/UserProfile3'
 
-import { GetMe } from '../../../../../../../apps/journeys-admin/__generated__/GetMe'
-import {
-  JourneyStatus,
-  Role
-} from '../../../../../../../apps/journeys-admin/__generated__/globalTypes'
-import { useAdminJourneysSuspenseQuery } from '../../../../../../../apps/journeys-admin/src/libs/useAdminJourneysSuspenseQuery'
-import { useUserRoleSuspenseQuery } from '../../../../../../../apps/journeys-admin/src/libs/useUserRoleSuspenseQuery'
-import { getJourneyTooltip } from '../../../../../../../apps/journeys-admin/src/components/PageWrapper/utils/getJourneyTooltip'
-import { JourneyTooltip } from '../../../../../../../apps/journeys-admin/src/components/PageWrapper/utils/getJourneyTooltip/getJourneyTooltip'
+import { JourneyStatus, Role } from '../../../../__generated__/globalTypes'
+import { getJourneyTooltip } from '../utils/getJourneyTooltip'
+import { JourneyTooltip } from '../utils/getJourneyTooltip/getJourneyTooltip'
+import { useAdminJourneysSuspenseQuery } from '../../../libs/useAdminJourneysSuspenseQuery'
+import { useUserRoleSuspenseQuery } from '../../../libs/useUserRoleSuspenseQuery'
+import { GetMe } from '../../../../../../libs/admin/ui/__generated__/GetMe'
 
 const ImpersonateDialog = dynamic(
   async () =>
