+++ libs/shared/ui/project.json
@@ -3,6 +3,7 @@
   "$schema": "../../../node_modules/nx/schemas/project-schema.json",
   "sourceRoot": "libs/shared/ui/src",
   "projectType": "library",
+  "implicitDependencies": ["locales"],
   "targets": {
     "lint": {
       "executor": "@nx/eslint:lint",
@@ -37,6 +38,16 @@
       "executor": "nx:run-commands",
       "options": {
         "command": "codecov -f coverage/libs/shared/ui/cobertura-coverage.xml -F libs.shared.ui"
+      }
+    },
+    "extract-translations": {
+      "executor": "@nx/workspace:run-commands",
+      "options": {
+        "commands": [
+          {
+            "command": "npx i18next --config PROJECT_PATH/i18next-parser.config.js"
+          }
+        ]
       }
     }
   },
